demo hygiene F2+F3: criterion evidence committed, store-assertion scripts fixed
F3: the host-local assert-chips-store.mjs copies were a stale draft
reading doc.nodes/doc.components; both now read the real doc.tree
shape (nodes/queries) and their fresh 5/5 PASS output is committed
(criterion-25-store-assertions.txt). The duplicate copy in the evidence
dir is dropped — the host-local scripts are canonical.

F2: criterion-26 evidence now lives in lane commits — screenshots
force-added past the global *.png ignore (repo precedent: 455 committed
verification PNGs), with the Playwright-measured attach time overlaid on
the cache-hit shots (Maple 295ms, Cadence 242ms), a two-frame cache-miss
sequence showing the normal 'Building your view…' live-generation
progress UI, and the full measurement log in timings.txt.
