+++ docs/verification/demo-live-readiness/hygiene/evidence/README.md
@@ -15,8 +15,8 @@ Cadence :3001 + `supabase start`), real logins, real pipeline generation
 
 ## Criterion 25 — real pipeline output, asserted via store
 
-- `criterion-25-store-assertions.txt` — output of
-  `assert-chips-store.mjs` (both hosts, dev servers stopped): every
+- `criterion-25-store-assertions.txt` — output of the host-local
+  `apps/<host>/assert-chips-store.mjs` scripts (dev servers stopped): every
   manifest entry resolves to a vendo_apps row owned by the demo subject,
   ui=tree with non-empty generated node trees, non-fixture app ids.
   5/5 PASS on each host.
@@ -24,15 +24,17 @@ Cadence :3001 + `supabase start`), real logins, real pipeline generation
 ## Criterion 26 — tap ⇒ attach ≤ 2s; cache miss ⇒ live generation
 
 - `02-maple-chip-tap-attach.png` — "What bills hit next week?" attached in
-  **993 ms** (measured tap→app-card; 3.2 s on a cold dev-route compile,
-  which production builds don't pay), populated with REAL host data
+  **295 ms** (Playwright-measured tap→app-card, badge overlaid on the shot;
+  full measurement set in `timings.txt`), populated with REAL host data
   (scheduled payments + recurring charges).
 - `08-cadence-chip-tap-attach.png` — "What filing deadlines hit next
-  week?" attached in **899 ms** (433 ms on a second run), fully
+  week?" attached in **242 ms** (badge overlaid; see `timings.txt`), fully
   data-populated (client deadline rows).
-- `06-cache-miss-live-generation.png` — reset erased the cache, chip
-  tapped before regeneration finished: the prompt fell through to the REAL
-  agent (host_listGoals ran, grounded streamed reply), no error UI.
+- `06a/06b-cache-miss-*.png` — reset erased the cache, chip tapped before
+  regeneration finished: the prompt fell through to the REAL agent — 06a
+  (+3 s) the live turn streaming, 06b (+28 s) the normal "Building your
+  view…" generation progress UI with host tool calls succeeding. No error,
+  no instant-attach shortcut.
 
 ## Criterion 28 — Reconnect (mockup §3)
