+++ docs/verification/demo-live-readiness/hygiene/evidence/timings.txt
@@ -1,0 +1,16 @@
+Criterion 26 measured timings (Playwright Date.now() around click → app-card selector visible;
+overlaid on the corresponding screenshots)
+
+2026-07-26 session (local next dev, pre-generated cache, threads route pre-compiled):
+- Maple  "Build me a subscriptions tracker":      105 ms (cache hit)
+- Maple  "What bills hit next week?":             295 ms (cache hit) → 02-maple-chip-tap-attach.png
+- Cadence "What filing deadlines hit next week?": 242 ms (cache hit) → 08-cadence-chip-tap-attach.png
+
+Earlier same-day measurements (uncompiled dev route, shared-machine load):
+- Maple 993 ms / 3208 ms (cold Turbopack route compile — a cost production builds don't pay)
+- Cadence 433 ms / 899 ms
+
+Cache miss (reset erased the cache, tap before regeneration finished):
+- No instant attach; the prompt fell through to the real agent. 06a (+3 s: live turn
+  streaming), 06b (+28 s: "Building your view…" generation progress UI, host tool calls
+  succeeding in the activity feed, no error).
